Fix schema migration for non-member nodes
This will likely fix the schema agreement issues we've been
experiencing. Schema needs to be migrated regardless of the node type.
The previous logic was only pulling the schema if the other node was a
stargate (non-member) node. That was incorrect schema should be pulled
regardless of the node type.