Move schema handling from DataStore to Persistence
`DataStore` was in charge of maintaining the schema view the persistence
API exposes. As a baby step toward refactoring `DataStore` to be solely
based on `Persistence`, this commit moves that schema handling to
`Persistence` (`DataStore` still has methods to access the schema, but
they now simply delegate to the underlying persistence).

This commit also remove `SchemaTool` (of persistence-commons): its one
`isInAgreement` method was using the Gossiper singleton, which always
linked to C* 3.11 (hard dependency of persistence-commons), which is
incorrect for any other persistence than the C* 3.11 one. The code
of that method has been inlined into the C* 3.11 persistence for now.