Cleanup of schema handling in persistence-api (#151)
* Move schema package from stargate.db.datastore to stargate.db

This is an automatic refactor. The rational is that the schema will
primarily move to `Persistence`, so won't be specific to datastore.

* Remove some dead code

* Add number coercion for values in ColumnType#validate

This commit also adds some Column type tests (for the validate method
and a few other existing ones).

* Move schema handling from DataStore to Persistence

`DataStore` was in charge of maintaining the schema view the persistence
API exposes. As a baby step toward refactoring `DataStore` to be solely
based on `Persistence`, this commit moves that schema handling to
`Persistence` (`DataStore` still has methods to access the schema, but
they now simply delegate to the underlying persistence).

This commit also remove `SchemaTool` (of persistence-commons): its one
`isInAgreement` method was using the Gossiper singleton, which always
linked to C* 3.11 (hard dependency of persistence-commons), which is
incorrect for any other persistence than the C* 3.11 one. The code
of that method has been inlined into the C* 3.11 persistence for now.

* Limit duplication of schema conversion code

This attempts to limit the amount of code duplication in persistence
implementation, mainly focusing on schema conversion as of this commit.

* Update code to check for schema agreement

First, the new method is more direct/shorter, so this is less code
duplication in all the persistence extensions in and of itself.

But also, the existing implementation has a few issues:
1. it was returning false (disagreement) as soon as an unreachable node
   wasn't in a "Gossip dead state". But "dead" in the sense of Gossip
   kinda means "is not part of the cluster", not "is down". So we
   would never call for agreement if any node was down, which does not
   feel like a good idea.
2. relying on basically string parsing of StorageProxy#getSchemaVersions
   feels a bit fragile.
3. that latter method works by querying all live nodes and waiting on
   all the answers. This makes it easy for that method to take a while
   (if a node doesn't answer, which makes the method wait until the
   RPC timeout), which isn't ideal given it's used to wait for
   agreement. Relying on the the gossip states, as this patch does,
   is faster and bypass such problems.

* Remove the parent link in InternalDataStore

It isn't used anymore now that the schema is in Persistence

Co-authored-by: Doug Wettlaufer <doug.wettlaufer@datastax.com>