Don't build iterable from stream with `#iterator`
Some call was building an iterable from a stream using the
`<stream>#iterator`. This is dodgy because such iterable can in fact
only be used once or will throw.

In this case, the `MaterializedView#create` method was iterating twice
on its `columns` argument which had been created just like that (once
to check a precondition, the 2nd time to truly use the columns).

The patch avoid this "pattern" by building a true iterable (a list) in
those cases.