Fix building of QueryOptions ignoring some options (#167)
Some of the native protocol query options (mainly names of bind markers
and client-provided timestamps) where not passed down properly to the
persistence layers.

And indeed, handling them requires reflection as the involved classes
and constructores are not currently public in C*/DSE.

This patch adds the required reflection magic to pass those options
properly. We also ensure the `QueryOptions#prepare` method is called
on all path, so named bound values are properly re-ordered.