Improve the handling of client warnings
In C*, warnings are handled through a thread local and use a singleton,
and Persistence was exposing that almost directly.  This isn't a very
good API however, so this patch changes it so warnings gets attached
to the result of queries (or, if the query errors out, to the exception
thrown).