Fix: Move CQL AuthResponse request handling to request threads
This moves AuthResponse request handling off of the event loop threads
to the appropriate request thread pool.

In C* (and DSE) the request logic normally runs on a separate threads
from the event loop threads. However, in Stargate, the logic for
AuthResponse is run directly on event loop threads which is unlike other
requests that run on Persistence backend threads. This was done so that
custom authentication logic could be added for Stargate. The problem
with running this logic on event loop threads is PasswordAuthenticator
runs BCrypt.checkpw(password, hash) which is quite CPU intensive
(intentionally) and causes significant delays to other requests.