Require authorization to read DML schemas
Motivation:

Unauthorized users were correctly prevented from executing data queries,
but they could still introspect the GraphQL schemas, which reveals
details about the CQL data model.

Modifications:

Perform an authorization check before we select the GraphQL schema to
use.

Result:

Unauthorized users will get a 401 for any GraphQL query, whether it's an
introspection query or one that goes to the backend.

Note that the DDL schema is not protected, it's a generic schema that is
the same for any Stargate installation, so exposing it doesn't reveal
any sensitive information.