Move the resetting of ExecutorLocals to the start of executeRequest
The pPrevious fix of resetting ExecutorLocals is not effective under
certain conditions (for example, when AuthCache uses RxThreads.singleFromCompletableFuture)
This may be a bug in DSE code, but as far as Stargate is concerned
resetting ExecutorLocals before building the Rx pipeline is probably
safer and more robust anyway.

DSE's TPCRunnable will take care of transferring thread local contexts
to the various RxJava tasks as needed (same as before).