Package integration tests as plain jar (#833)
This is to allow clients to control dependencies in their build env.

* Move test code into the main/java dir
* Drop the fat testing*-all.jar

Note: this will make using JUnit5 console launcher more complicated,
but it is not expected to be a common use case. Downstream projects
usign Maven or other build tools should still be able to automatically
run all tests with their respective dependent libs (and also control
dependency versions).