Support paginating over nested doc elements
This is now enabled when the page-size parameter is used even without a where clause.

Note: this is a slight change in API behaviour as previously
the page-size parameter was ignored in this case. However, there
was no reason for specifying the page size in this case before,
so hopefully end-user use cases are not broken by this change.

Fixes #1040