Add a gRPC interceptor that caches `Connection`
Profiling found that validating the token using the authentication
service and constructing a new `Connection` object was consuming a
significant percentage of a request's time.

This adds a new interceptor called `NewConnectionInterceptor` which
combines the functionality of the authentication, headers, and remote
address interceptor and caches the resulting `Connection` object.