Add a gRPC interceptor that caches `Connection` (#1241)
* Add a gRPC interceptor that caches `Connection`

Profiling found that validating the token using the authentication
service and constructing a new `Connection` object was consuming a
significant percentage of a request's time.

This adds a new interceptor called `NewConnectionInterceptor` which
combines the functionality of the authentication, headers, and remote
address interceptor and caches the resulting `Connection` object.

* Prevent `USE <keyspace>` from being executed

Caching the connection allows the keyspace be set in the `ClientState`
between requests and allows there to be a default keyspace. The gRPC API
is expected to be stateless and this is unexpected behavior.