Encode Udt field values in the correct order
UdtCodec#encode was encoding the fields in the order
of values received from the `UserDefinedType#columnMap`.
That order does not necessarily match the order of the fields
in the UDT, because the returned map is a HashMap.
Placing the values in the output buffer in a wrong order can
lead to silent data corruption or at best
marshalling exceptions from C*.

This commit makes the encoded values order match exactly
the order of fields in the UDT.

Fixes #1329