Fix gRPC prepared cache invalidation approach #2
This approach fixes prepared cache validation by registering an
`EventListener` to invalidate external prepared cache whenever relevant
schema change event occur. This means that the cache can no longer be
asynchronous and will possibly suffer from a prepare stampede when the
same query is run many times simultaneously.

The drawback of this approach is the very rough invalidation of the
cache and complexity. The invalidation of the cache can be quite broad
i.e. if we're unable to determine the keyspace and/or table of a
prepared statement we have to invalidate it because we don't have enough
metadata to see if it's still valid or not.