Fix gRPC prepared cache invalidation approach #1 (#1335)
This approach fixes prepared cache invalidation by using the internal
cache only. Cassandra backends don't switch to another thread to handle
prepared statements so they don't require any special implementation.
DSE uses a TPC thread so it provides a implementation that attempts to
access the internal prepared cache directly first before running a
prepare on the TPC thread.

The disadvantage of of this approach is constructing the
`Result.Prepared` is pretty heavy and caching in the gRPC layer also
caches that work. Also, we need to add yet another method to the
persistence interface which is not great...