gRPC: Set "host" header to ":authority" pseudo-header value
When `grpc-netty-shaded` was upgraded to `1.42.1` it included this change:

https://github.com/grpc/grpc-java/commit/122b3b2f7cf2b50fe0a0cebc55a84133441a4348

That change removes the "host" header from the request header metadata.
The problem is internal authentication and other persistence-level
logic depend on the "host" header being present.

This patch sets the internal "host" header using the ":authority"
pseudo-header value.

Additionally, this adds logging to the unhandled error path when an
exception is thrown while creating a new connection.