Create dependabot.yml (#2530)
I've been looking at our current approach using a homegrown workflow running maven commands to track dependency updates and vulnerabilities. After reading the dependabot docs in more detail I think it has the potential to be a better solution, with support for checking more than just Java dependencies, and better automation to help us keep current more quickly. I'd like to give it a shot on this repo and then see if we want to also enable on other repos.